Create ExternalUIDManager to incorporate external UID logic
